split the document
